refactor bi arm task.
