+++ source/leisaac/leisaac/tasks/template/__init__.py
@@ -1,1 +1,2 @@
 from .single_arm_env_cfg import SingleArmTaskSceneCfg, SingleArmTaskEnvCfg, SingleArmObservationsCfg, SingleArmTerminationsCfg
+from .bi_arm_env_cfg import BiArmTaskSceneCfg, BiArmTaskEnvCfg, BiArmObservationsCfg, BiArmTerminationsCfg
